% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_data.R
\docType{data}
\name{annot_data}
\alias{annot_data}
\title{List contains annotation data}
\format{A list with 5 entries
\describe{
    \item{annot_data}{list contains GO term ids as keys and corresponding sets of genes as values}
    \item{relation_data}{list contains GO term ids as keys and parent term as value}
    \item{real_names}{list contains GO term ids as keys and GO terms real names as values}
    \item{namespaces}{list contains GO term ids as keys and GO terms namespaces as values}
    \item{annotation_metadata}{annotation metadata}
}}
\usage{
annot_data
}
\description{
List contains annotation data
}
\keyword{datasets}
