% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_genes.R
\docType{data}
\name{bg_genes}
\alias{bg_genes}
\title{Vector contains background set of genes from example rna-seq experiment data}
\format{A dataframe with 4 columns}
\usage{
bg_genes
}
\description{
Vector contains background set of genes from example rna-seq experiment data
}
\keyword{datasets}
