% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_data_to_wide.R
\name{convert_annotation_wide}
\alias{convert_annotation_wide}
\title{Convert GAF format type annotation to list contains GO term id's as keys and Gene ID's as values}
\usage{
convert_annotation_wide(data, id_pattern = NULL)
}
\arguments{
\item{data}{- GAF annotation table (output of read.gaf function)}

\item{id_pattern}{- regex for Gene identifier (see fsgor::patterns_enum$Arabidopsis_thaliana for Arabidopsis AGI code example pattern)}
}
\value{
- list with GO term id's as keys and Gene ID's as values
}
\description{
Convert GAF format type annotation to list contains GO term id's as keys and Gene ID's as values
}
\examples{
\dontrun{
    convert_annotation_wide(annot_data, fsgor::patterns_enum$Arabidopsis_thaliana)
}
}
