% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ancestry_annot.R
\name{create_annotation_with_ancestry}
\alias{create_annotation_with_ancestry}
\title{Annotation with ancestry creator}
\usage{
create_annotation_with_ancestry(rellist, annotlist)
}
\arguments{
\item{rellist}{list contains child terms as keys and nearest parents as values}

\item{annotlist}{list contains terms as keys and direct annotations as values}
}
\value{
list contains terms as keys and genes annotated according to True Path Rule as values
}
\description{
Annotaion for terms taking into account True Path Rule
}
\examples{
\dontrun{
    onto_data <- fsgor::read_obo("~/go.obo")
    annot_data <- fsgor::read_gaf("~/goa_human.gaf")
    annot_data <- fsgor::convert_annotation_wide(annot_data$data)
    create_annotation_with_ancestry(onto_data$relation_data, annot_data)
}

}
