% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentiler.R
\name{div_genes_to_quantiles}
\alias{div_genes_to_quantiles}
\title{Gene set quantile separator}
\usage{
div_genes_to_quantiles(data, n)
}
\arguments{
\item{data}{dataframe contains initial set of genes gene ID's in the first row
and corresponding fold change values in the second row}

\item{n}{number of quantiles}
}
\value{
list of gene sets for each quatile and all combinations
}
\description{
Converts the set of genes in to quantiles according log value of fold change and
produces all possible combinations of resulting gene subsets.
Input dataset must be dataframe with gene ID's in the first row and corresponding
fold change values in the second row. Note: dataframe must contain only genes with
common type of regulation (up or down)
}
\examples{
    data(up_genes, package = "fsgor")
    gene_sets <- div_genes_to_quantiles(up_genes, 6)

}
