% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_genes.R
\docType{data}
\name{down_genes}
\alias{down_genes}
\title{Dataframe contains example data for div_genes_quantiles function}
\format{A dataframe with 4 columns
\describe{
    \item{V1}{Gene ID}
    \item{V2}{fold-change value}
    \item{V3}{p-value}
    \item{V4}{adjusted p-value}
}}
\usage{
down_genes
}
\description{
Dataframe contains example data for div_genes_quantiles function
}
\keyword{datasets}
