% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_rectangle_plot.R
\name{fold_spec_chart}
\alias{fold_spec_chart}
\title{fold specificity rectangle plot}
\usage{
fold_spec_chart(fs_res_up, fs_res_down, scope = NULL, interval_labels,
  x_text_size = 10)
}
\arguments{
\item{fs_res_up}{dataframe with fold-specific GO terms data (recognize_fs_terms() function output) for up regulated genes}

\item{fs_res_down}{dataframe with fold-specific GO terms data (recognize_fs_terms() function output) for down regulated genes}

\item{scope}{GO term namespace ("biological_process", "molecular_function", "cellular_component")}

\item{interval_labels}{vector of user defined names for non-overlaping intervals (e.g. c("weak response",...,"strong response")), see fsgor::six_perc_labs for example}

\item{x_text_size}{size of text for x axis labels}
}
\value{
fold specificity rectangle plot as ggplot object
}
\description{
Create rectangle plot for fold specificity data representation
}
\examples{
    data(fs_res_up, fs_res_down, package = "fsgor")
    fold_spec_chart(fs_res_up, fs_res_down, interval_labels = fsgor::six_perc_labs)
}
