% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_data_viz_preparation.R
\name{fold_spec_chart_data}
\alias{fold_spec_chart_data}
\title{Data parser for fold-specificity rectangle plot}
\usage{
fold_spec_chart_data(fs_res_up, fs_res_down, scope = NULL)
}
\arguments{
\item{fs_res_up}{dataframe contains fold-specificty recognition data (recognize_fs_terms function output) for up regulation}

\item{fs_res_down}{dataframe contains fold-specificty recognition data (recognize_fs_terms function output) for down regulation}

\item{scope}{namespace of terms for plot input data (NULL by default)}
}
\value{
input dataframe for create.fold.spec.chart function
}
\description{
Create input data for rectangle plot (fold_spec_chart function) using recognize_fs_terms function output as input
}
\examples{
    data(fs_res_up, fs_res_down, package = "fsgor")
    fold_spec_chart_data(fs_res_up, fs_res_down, scope = "molecular_function")
}
