% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_spec_graph.R
\name{fold_spec_graph}
\alias{fold_spec_graph}
\title{Create igraph object for fold-specificity recognition function output (recognize_fs_terms())}
\usage{
fold_spec_graph(fs_res, annot_data, reg, scope = NULL)
}
\arguments{
\item{fs_res}{- list contains fold-specificity recognition results}

\item{annot_data}{- prepare_annotation() function output}

\item{reg}{- regulation type for DEG taken in to analysis ("up" or "down")}

\item{scope}{- GO term namespace ("biological_process", "cellular_component", "molecular_function"),
NULL by default (GO terms from all namespaces will be plotted)}
}
\value{
- igraph object with sugiyama layout
}
\description{
Create igraph object for fold-specificity recognition function output (recognize_fs_terms())
}
\examples{
     data(fs_res_up, annot_data, package = "fsgor")
     fold_spec_graph(fs_res_up, annot_data, "up", scope = "molecular_function")
}
