% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_res_up.R
\docType{data}
\name{fs_res_up}
\alias{fs_res_up}
\title{List contains dataframes with fold-specificty recognition results for example set of up-regualted genes}
\format{A list with 2 entries
\describe{
    \item{fold-change-specific GO terms}{dataframe consitst of following columns: ids, namespace, name, padj, interval, genes}
    \item{not fold-change-specific GO terms}{dataframe consitst of following columns: ids, namespace, name, padj, interval, genes}
}}
\usage{
fs_res_up
}
\description{
List contains dataframes with fold-specificty recognition results for example set of up-regualted genes
}
\keyword{datasets}
