% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_annot_test.R
\name{func_annot_test_internal}
\alias{func_annot_test_internal}
\title{Functional annotation test}
\usage{
func_annot_test_internal(annotlist, realnames, namespaces, inputgeneset,
  gene_amount_border = 1, fisher_alternative = "greater",
  p_adjust_method = "BH")
}
\arguments{
\item{annotlist}{- list contains terms as keys and direct annotations as values}

\item{realnames}{- GO terms full names}

\item{namespaces}{- GO terms namespaces}

\item{inputgeneset}{- vector of query genes}

\item{gene_amount_border}{- minimal number of genes annotated to a term (1 by default)}

\item{fisher_alternative}{- fisher exact test alternative,
possible values: "greater", "less", "two.sided" ("greater by default)}

\item{p_adjust_method}{- method for multiple testing correction (to see all possible methods print: p.adjust.methods)}
}
\value{
- dataframe contains functional annotaion for gene set
}
\description{
Conducts functional annotation using Fisher exact test
}
\examples{
    # data - dataframe contains gene IDs in first column and fold-change values in second column
    data(up_genes, annot_data, package = "fsgor")
    gene_sets <- div_genes_to_quantiles(up_genes, 6)
    func_annot_test_internal(annot_data$annot_data, annot_data$real_names, annot_data$namespaces, gene_sets$'1')
}
