% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_spec_graph.R
\name{get_term_ancestors}
\alias{get_term_ancestors}
\title{Extract all ancestors for distinct GO term}
\usage{
get_term_ancestors(term_vec, relation_data)
}
\arguments{
\item{term_vec}{- vector of terms}

\item{relation_data}{- ontology relation data (prepare_annotation(...)$relation_data}
}
\value{

}
\description{
Extract all ancestors for distinct GO term
}
\examples{
\dontrun{
    onto_data <- fsgor::read_obo("~/go.obo")
    get_term_ancestors(term_vec, onto_data$relation_data)
}
}
