% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_spec_graph.R
\name{igraph_data_prep}
\alias{igraph_data_prep}
\title{Create igraph object for fold-specific terms}
\usage{
igraph_data_prep(fs_terms, relation_data, reg)
}
\arguments{
\item{fs_terms}{- vector contains fold-specific terms GO ids}

\item{relation_data}{- ontology relation data (prepare_annotation(...)$relation_data}

\item{reg}{- regulation type for DEG taken in to analysis ("up" or "down")}
}
\value{
- igraph object
}
\description{
Create igraph object for fold-specific terms
}
\examples{
\dontrun{
   data(fs_res_up, annot_data)
   igraph_data_prep(fs_res_up$fs$ids, annot_data, "up")
}
}
