% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_with_background.R
\name{intersect_with_background}
\alias{intersect_with_background}
\title{Intersect annotation with background set of genes}
\usage{
intersect_with_background(annot_data, background_genes)
}
\arguments{
\item{annot_data}{list contains terms as keys and direct annotations as values}

\item{background_genes}{vector contains background set of genes}
}
\value{
list contains terms as keys and direct annotations as values with genes only from background
}
\description{
Intersect list contains terms as keys and direct annotations as values
with vector contains background set of genes
}
\examples{
    data(annot_data, bg_genes, package = "fsgor")
    intersect_with_background(annot$annot_data, bg_genes)
}
