% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_sets_annot.R
\name{multiple_sets_annot}
\alias{multiple_sets_annot}
\title{Functional annotation for multiple gene sets}
\usage{
multiple_sets_annot(gene_sets, annot_data, gene_amount_border = 1,
  background_genes = NULL, p_adjust_method = "BH")
}
\arguments{
\item{gene_sets}{list contains quantile names (e.g. "1","2","1-6") as keys and corresponding gene sets as values}

\item{annot_data}{list object contains annotation data ready for functional annnotation procedure,
ontology relation data, GO terms full names, GO terms namespaces and metadata for annotation}

\item{gene_amount_border}{minimal number of genes annotated to a term (1 by default)}

\item{background_genes}{vector contains background set of genes}

\item{p_adjust_method}{method for multiple testing correction (to see all possible methods print: p.adjust.methods)
Benjamini-Hochberg by default}
}
\value{
list with filenames as keys and annotaton data frames as values
}
\description{
Conducts functional annotation using Fisher exact test
for multiple gene sets, returns list with resulting annotations
}
\examples{
    data(up_genes, annot_data, package = "fsgor")
    gene_sets <- div_genes_to_quantiles(up_genes, 6)
    multiple_sets_annot(gene_sets, annot_data,  gene_amount_border = 10)
}
