% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_enum.R
\docType{data}
\name{patterns_enum}
\alias{patterns_enum}
\title{List contains regex patterns for Gene identifiers}
\format{A list with 1 entry
\describe{
    \item{Arabidopsis_thliana}{regex pattern for Arabidopsis thaliana gene identifier}
}}
\usage{
patterns_enum
}
\description{
List contains regex patterns for Gene identifiers
}
\keyword{datasets}
