% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation.R
\name{prepare_annotaion}
\alias{prepare_annotaion}
\title{Create annotation object from .obo and .gaf files}
\usage{
prepare_annotaion(onto_path, annot_path)
}
\arguments{
\item{onto_path}{- full path to .obo file}

\item{annot_path}{- full path to .gaf file}
}
\value{
- list object contains annotation data ready for functional annnotation procedure,
          ontology relation data, GO terms full names, GO terms namespaces and metadata for annotation
}
\description{
Create annotation object from .obo and .gaf files
}
\examples{
    onto_path <- system.file("extdata", "goslim_plant.obo", package = "fsgor")
    annot_path <- system.file("extdata", "gene_association.tair", package = "fsgor")
    annot <- prepare_annotation(onto_path, annot_path)

    # extract annotation calculated considering GO term ancestry
    annot$annot_data

    # GO term ontology relations
    annot$relation_data

    # GO terms real names
    annot$real_names

    # GO terms namespaces
    annot$namespaces

    # annotation metadata
    annot$annot_metadata

}
