% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAF_reader.R
\name{read_gaf}
\alias{read_gaf}
\title{GAF file reader}
\usage{
read_gaf(gafpath)
}
\arguments{
\item{gafpath}{- full path to GAF file}
}
\value{
- list contains information about file and data accessible by $info and $data keys correspondingly
}
\description{
GAF file reader
}
\examples{
    gaf_path <- system.file("extdata", "gene_association.tair", package = "fsgor")
    read_gaf(gaf_path)
}
