% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBO_reader.R
\name{read_obo}
\alias{read_obo}
\title{OBO file reader}
\usage{
read_obo(obopath)
}
\arguments{
\item{obopath}{path to the .obo format file}
}
\value{
lists that contain GO terms ancestry data, names for GO terms, and corresponding namespace for each GO term
}
\description{
Parses .obo format file (open biomedical ontologies), returns list contains GO term ontology relation data,
GO terms real names, GO terms namespaces
}
\examples{
    go_path <- system.file("extdata", "goslim_plant.obo", package = "fsgor")
    read_obo(go_path)
}
