% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_spec_recognition.R
\name{recognize_fs_terms}
\alias{recognize_fs_terms}
\title{Fold-specific GO terms recognition}
\usage{
recognize_fs_terms(listoftables, wholeintname,
  listofcolnames = fsgor::colnameslist, p_adjust_method, fdr2step, fdr3step,
  fisher_alternative = "greater")
}
\arguments{
\item{listoftables}{list contains GO output tables as dataframes and interval names as leys}

\item{wholeintname}{key for whole interval table specified in listoftables}

\item{listofcolnames}{list of column names, must contain the following keys: "GO_id" - GO term identifier
"name" - GO term name
"namespace" - namespace of GO term
"qit" - number of genes annotated to a term in query list
"qtot" - number of genes in a query list
"padj" - adjusted p-values, if you don't use any corrections on multiple testing assign the name of raw p-values column to this key
(column names of internal annotation used as default value)}

\item{p_adjust_method}{method name for correction on multiple correction, type "p.adjust.methods" in R Console
to obtain all possiple values}

\item{fdr2step}{value for fdr step 2}

\item{fdr3step}{value for fdr step 3}

\item{fisher_alternative}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. Only used in the 2 by 2 case.}
}
\value{
list contains dataframe with fold-specific and not fold-fold-specific GO terms data
}
\description{
Recognizes fold-specific GO for each fold change interval versus differentialy expressed genes annotation,
returns list contains dataframe with fold-specific GO terms data and not fold-fold-specific GO terms data
accessible by fs and nfs keys correspondingly
}
\examples{
    data(up_annot, package = "fsgor")
    recognize_fs_terms(up_annot, "1-6", fsgor::colnameslist, "BY", 1, 0.05)
}
