% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/six_perc_labs.R
\docType{data}
\name{six_perc_labs}
\alias{six_perc_labs}
\title{Labels for fold_spec_chart with data separated into six quantiles}
\format{A vector contains 6 characters}
\usage{
six_perc_labs
}
\description{
Labels for fold_spec_chart with data separated into six quantiles
}
\keyword{datasets}
