% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up_annot.R
\docType{data}
\name{up_annot}
\alias{up_annot}
\title{List contains functional annotations for all combinatons of quantiles derived from initial example gene set}
\format{A list with 21 entries
\describe{
    \item{quantile number}{dataframe consitst of following columns: GO_id, namespace, name, qit, qtot, bgit, bgtot, pval, padj}
}}
\usage{
up_annot
}
\description{
List contains functional annotations for all combinatons of quantiles derived from initial example gene set
}
\keyword{datasets}
